+++ cvat-cli/src/cvat_cli/core/__init__.py
@@ -1,3 +1,6 @@
+# Copyright (C) 2020-2022 Intel Corporation
+#
 # SPDX-License-Identifier: MIT
+
 from .definition import parser, ResourceType  # noqa
 from .core import CLI, CVAT_API_V2  # noqa