+++ cvat-cli/tests/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (C) 2020-2022 Intel Corporation
+#
+# SPDX-License-Identifier: MIT