+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Project/task backups uploading via chunk uploads (<https://github.com/cvat-ai/cvat/pull/9>)
 - Fixed UX bug when jobs pagination is reset after changing a job (<https://github.com/cvat-ai/cvat/pull/42>)
 - Progressbars in CLI for file uploading and downloading (<https://github.com/cvat-ai/cvat/pull/46>)
+- `utils/cli` changed to `cvat-cli` package (<https://github.com/cvat-ai/cvat/pull/59>)
 
 ### Changed
 - Bumped nuclio version to 1.8.14 (<https://github.com/cvat-ai/cvat/pull/29>)
@@ -25,8 +26,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Task creation progressbar bug (<https://github.com/cvat-ai/cvat/issues/12>)
 - Removed Python dependency ``open3d`` which brought different issues to the building process
 - Analytics not accessible when https is enabled (<https://github.com/cvat-ai/cvat/pull/20>)
-- Dataset import in an organization (<https://github.com/cvat-ai/cvat/pull/18>)
+- Dataset import in an organization (<https://github.com/cvat-ai/cvat/pull/18>, <https://github.com/cvat-ai/cvat/pull/35>)
 - Updated minimist npm package to v1.2.6 (<https://github.com/cvat-ai/cvat/pull/34>)
+- Request Status Code 500 "StopIteration" when exporting dataset (<https://github.com/cvat-ai/cvat/pull/17>)
 - Generated OpenAPI schema for several endpoints (<https://github.com/cvat-ai/cvat/issues/72>)
 
 ### Security